moved conduct_chain mod to a file
